feat: extract code input component
